Override old symlinks in setup-config.sh

+++ scripts/setup-config.sh
@@ -16,23 +16,15 @@ REPO_DIR="$( dirname "${SCRIPT_DIR}" )"
 NGINX_CONF_DIR="/etc/nginx"
 NGINX_CONF="${NGINX_CONF_DIR}/nginx.conf"
 NGINX_CONF_D_DIR="${NGINX_CONF_DIR}/conf.d"
-if [ -L "${NGINX_CONF_D_DIR}" ]; then
-  echo "conf.d link already exists."
-else
-  if [ -d "${NGINX_CONF_D_DIR}" ]; then
-    mv "${NGINX_CONF_D_DIR}" "${NGINX_CONF_D_DIR}-old"
-  fi
-  ln -s "${REPO_DIR}/config/conf.d" "${NGINX_CONF_D_DIR}"
+if [ -d "${NGINX_CONF_D_DIR}" ] && [ ! -L "${NGINX_CONF_D_DIR}" ]; then
+  mv "${NGINX_CONF_D_DIR}" "${NGINX_CONF_D_DIR}-old"
 fi
+ln -f -s "${REPO_DIR}/config/conf.d" "${NGINX_CONF_D_DIR}"
 
-if [ -L "${NGINX_CONF}" ]; then
-  echo "nginx.conf link already exists."
-else
-  if [ -f "${NGINX_CONF}" ]; then
-    mv "${NGINX_CONF}" "${NGINX_CONF}-old"
-  fi
-  ln -s "${REPO_DIR}/config/nginx-ubuntu.conf" "${NGINX_CONF}"
+if [ -f "${NGINX_CONF}" ] && [ ! -L "${NGINX_CONF}" ]; then
+  mv "${NGINX_CONF}" "${NGINX_CONF}-old"
 fi
+ln -f -s "${REPO_DIR}/config/nginx-ubuntu.conf" "${NGINX_CONF}"
 
 if [ -f "${NGINX_CONF_DIR}/sites-enabled/default" ]; then
   rm "${NGINX_CONF_DIR}/sites-enabled/default"
